LayoutTests/media/video-muted-expected.txt

TEST(!video.muted) OK
TEST(video.muted) OK
TEST(!video.muted) OK
END OF TEST
